+++ spec/lib/karafka/controllers/includer_spec.rb
@@ -10,16 +10,16 @@ RSpec.describe Karafka::Controllers::Includer do
     instance_double(
       Karafka::Routing::Topic,
       backend: backend,
-      batch_processing: batch_processing,
+      batch_consuming: batch_consuming,
       responder: responder
     )
   end
 
   before { controller_class.topic = topic }
 
-  describe 'inline with batch processing' do
+  describe 'inline with batch consuming' do
     let(:backend) { :inline }
-    let(:batch_processing) { true }
+    let(:batch_consuming) { true }
     let(:responder) { nil }
 
     it { expect(controller_class.include?(backends_scope::Inline)).to eq true }
@@ -27,9 +27,9 @@ RSpec.describe Karafka::Controllers::Includer do
     it { expect(controller_class.include?(features_scope::Responders)).to eq false }
   end
 
-  describe 'inline without batch processing' do
+  describe 'inline without batch consuming' do
     let(:backend) { :inline }
-    let(:batch_processing) { false }
+    let(:batch_consuming) { false }
     let(:responder) { nil }
 
     it { expect(controller_class.include?(backends_scope::Inline)).to eq true }
@@ -39,7 +39,7 @@ RSpec.describe Karafka::Controllers::Includer do
 
   describe 'inline with responder' do
     let(:backend) { :inline }
-    let(:batch_processing) { false }
+    let(:batch_consuming) { false }
     let(:responder) { Class.new }
 
     it { expect(controller_class.include?(backends_scope::Inline)).to eq true }