+++ spec/lib/karafka/schemas/consumer_group_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe Karafka::Schemas::ConsumerGroup do
         parser: Class.new,
         max_bytes_per_partition: 1,
         start_from_beginning: true,
-        batch_processing: true,
+        batch_consuming: true,
         persistent: false
       }
     ]